s3: Fix some 64-bit warnings

+++ source3/rpc_client/cli_pipe.c
@@ -976,7 +976,9 @@ static NTSTATUS cli_pipe_validate_current_pdu(TALLOC_CTX *mem_ctx,
 		}
 
 		DEBUG(10, ("Got pdu len %lu, data_len %lu, ss_len %u\n",
-			   pdu->length, rdata->length, ss_padding_len));
+			   (long unsigned int)pdu->length,
+			   (long unsigned int)rdata->length,
+			   (long unsigned int)ss_padding_len));
 
 		/*
 		 * If this is the first reply, and the allocation hint is
